+++ SuperMarioBros3/GameMap.h
@@ -55,10 +55,10 @@ class CChimneyLayout
 {
 public:
 	int id;
-	int x;
-	int y;
+	float x;
+	float y;
 
-	CChimneyLayout(int id, int x, int y) {
+	CChimneyLayout(int id, float x, float y) {
 		this->id = id;
 		this->x = x;
 		this->y = y;
